requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
streamlit>=1.22.0
plotly>=5.13.0
python-dotenv>=1.0.0 